refactor: share global ID encoding
